requirements.txt
allure_python_commons==2.12.0
colorlog==4.8.0
ddddocr==1.4.7
openpyxl==3.1.2
Pillow==9.5.0
Pillow==10.0.1
playwright==1.28.0
pytest==7.2.0
PyYAML==6.0
PyYAML==6.0.1
